Event Popup filtering properly
